Just send Notifications without an Intent
I haven’t been able to remove geofences when the app is closed, and as
a result, I haven’t been able to prevent GeofenceIntentService from
sending notifications when the user enters a geofence, when the app is
closed.

The idea behind the notification was to bring the map (and all its
geofence markers) to the foreground, when the notification was clicked.
However, this map is not initialized when the app is closed, but the
geofences are still active (per the issue described above). When a user
clicks the notification in this scenario, Cortez will crash.

Since we still want to send notifications but I’m not sure if I’ll have
time to fix this bug, I’m removing the ability for the notification to
take the user to the map.